Add PHP8 Support
PHP 8 is the current active release and PHP 7 entered the security support, so
new versions should support PHP 8 from now on.
Dependencies were locked to the most up-to-date version and suggestions
were changed based on this change as well.

DateTimeProcessor class was removed, because it seems unnecessary and it was
breaking with new versions of monolog (2.3.0+). Current datetime is required to
be a DateTimeImuttable instance and the default implementations already uses
the RFC 3339 format.